microservices/notification_service/requirements.txt
# UK Management Bot - Notification Service Dependencies

# FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Redis for caching and events
redis[hiredis]==5.0.1
aioredis==2.0.1

# Telegram Bot API
aiogram==3.2.0

# Authentication
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client
httpx==0.25.2
requests==2.31.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# OpenTelemetry tracing
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp==1.21.0

# Prometheus metrics
prometheus-client==0.19.0

# Validation and serialization
marshmallow==3.20.2
pydantic[email]==2.5.0

# Utilities
tenacity==8.2.3
faker==20.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Development
black==23.11.0
ruff==0.1.6
mypy==1.7.1

# Logging
structlog==23.2.0